requirements.txt
# --------- pytorch --------- #
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision
torchmetrics
torchinfo
numpy

# --------- quantization --------- #
--extra-index-url https://pypi.ngc.nvidia.com
pytorch-quantization
#-f https://github.com/NVIDIA/Torch-TensorRT/releases
#torch-tensorrt
brevitas

# --------- hydra --------- #
hydra-core==1.1.1
hydra-submitit-launcher

# --------- pytest --------- #
pytest

# --------- matplotlib --------- #
matplotlib

# --------- tqdm --------- #
tqdm

# --------- filesystem handling --------- #
pathlib
pandas
pyyaml
